refactor(MenuInitiator): 패키지 이동
-> utils
