+++ src/main/java/menu/utils/MenuInitiator.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.utils;
 
 import java.util.Arrays;
 import java.util.List;
